+++ src/app/cart/components/cart-list/cart-list.component.spec.ts
@@ -1,6 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { CartListComponent } from './cart-list.component';
+import { CartModule } from '../../cart.module';
 
 describe('CartListComponent', () => {
   let component: CartListComponent;
@@ -8,6 +9,7 @@ describe('CartListComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
+      imports: [ CartModule ],
       declarations: [ CartListComponent ]
     })
     .compileComponents();