+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -8,16 +8,28 @@ import { ProductModel } from 'src/app/products/models/product-model';
   styleUrls: ['./cart-list.component.css']
 })
 export class CartListComponent {
+  
   constructor(private cartService: CartService) {}
 
-  getProducts(): ProductModel[] {
+  onQuantityIncrease(product: ProductModel) {
+    this.cartService.addProduct(product)
+  }
+  onDeleteItem(product: ProductModel) {
+    this.cartService.removeProduct(product, true)
+  }
+  onQuantityDecrease(product: ProductModel) {
+    this.cartService.removeProduct(product)
+  }
+
+  getItems(): {product: ProductModel, quantity: number}[] {
     return this.cartService.getProducts()
   }
-
-  getSumPrice(): Number {
-    return this.cartService.getSumPrice()
+  getTotalCost(): Number {
+    return this.cartService.getTotalCost()
   }
-
+  getTotalQuantity(): Number {
+    return this.cartService.getTotalQuantity()
+  }
   isEmpty(): boolean {
     return this.cartService.isEmpty()
   }