+++ src/app/cart/models/cart-model.ts
@@ -1,5 +1,10 @@
 import { ProductModel } from "src/app/products/models/product-model";
 
+export type CartItem = {
+    product: ProductModel,
+    quantity: number
+}
+
 export interface CartModel {
-    products: ProductModel[];
+    items: CartItem[]
 }